Various consistency fixes; semconv improvements

+++ common/telemetry/metrics/instruments/RdbmsInstruments.hpp
@@ -21,7 +21,7 @@
 
 namespace cta::telemetry::metrics {
 
-extern std::unique_ptr<opentelemetry::metrics::Histogram<double>> rdbmsOperationDuration;
+extern std::unique_ptr<opentelemetry::metrics::Histogram<uint64_t>> rdbmsOperationDuration;
 // TODO: add attributes to the recordings here
 extern std::unique_ptr<opentelemetry::metrics::UpDownCounter<int64_t>> rdbmsConnectionCount;
